refactor: constants 경로 수정
